requirements.txt
python-telegram-bot==21.7
openai==0.28.1
python-dotenv==1.0.1
tqdm==4.66.1
psycopg==3.1.18
boto3==1.34.0  # For optional S3 uploads